+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - **New Feature**
   - make Array<T> a HKT and deprecate `to`,`from` helper functions, fix #5 (@gcanti)
   - add `Traced` comonad (@bumbleblym)
+  - add `getOrElse` method to `Option` (@gcanti)
 - **Polish**
   - add tslint